Unicode.Grapheme

Description

Text grapheme utilities.

Since: 0.1

Documentation

Unicode functions are defined in terms of the abstract UnicodeFunction type, which allows us to conveniently wrap functionality across multiple unicode versions.

These can then be combined in a variety of ways for handling the unicode version.

For example, the following function will break the text into grapheme clusters, using either base's unicode version if it is supported, or falling back to the latest supported version.

  break :: Text -> [Text]
  break = runUnicodeFunction breakGraphemeClusters

data UnicodeFunction a b

UnicodeFunction represents some function that works across all UnicodeVersions.

Since: 0.1

breakGraphemeClusters :: UnicodeFunction Text [Text]

Breaks Text into grapheme clusters.

Since: 0.1

textWidth :: UnicodeFunction Text Int

Splits the text into grapheme clusters and counts each cluster width.

Since: 0.1

clusterWidth :: UnicodeFunction Text Int

Given a single grapheme cluster, returns the width 1 or 2. This is based on heuristics i.e. if the text contains at least one codepoint with the following properties:

East_Asian_Width = Fullwidth or Wide

Emoji_Presentation

U+FE0F (emoji-style)

Then width is 2. Otherwise it is 1.

Since: 0.1

Version Combinators

runUnicodeFunction :: UnicodeFunction a b -> a -> b

Runs the UnicodeFunction with base's unicode version, if it is supported. Otherwise uses the latest supported version.

Since: 0.1

runUnicodeFunctionVersion :: UnicodeVersion -> UnicodeFunction a b -> a -> b

Runs the UnicodeFunction with the given unicode version.

Since: 0.1

Unicode versions

data UnicodeVersion

Supported unicode versions. The following table lists the unicode versions for each base that is usable with this library. The Supported column refers to if the unicode version itself is supported here.

Base	Unicode	Supported

4.21	16.0	🌕

4.20	15.1	🌕

4.19	15.1	🌕

4.18	15.0	🌕

4.17	14.0	🌑

4.16	14.0	🌑

Since: unicode-grapheme-common-0.1

Constructors

UnicodeVersion_15_0
Since: unicode-grapheme-common-0.1

UnicodeVersion_15_1
Since: unicode-grapheme-common-0.1

UnicodeVersion_16_0
Since: unicode-grapheme-common-0.1

Functions

getBaseUnicodeVersion :: Either UnsupportedUnicodeE UnicodeVersion

Retrieves base's unicode version, or an error if it is unsupported.

Since: unicode-grapheme-common-0.1

getBaseUnicodeVersionIO :: IO UnicodeVersion

Retrieves base's unicode version or throws UnsupportedUnicodeE.

Since: unicode-grapheme-common-0.1

getBaseUnicodeVersionOrLatest :: UnicodeVersion

Retrieves base's unicode version or the latest, if the former is unsupported.

Since: unicode-grapheme-common-0.1

Display

displayVersion :: IsString s => UnicodeVersion -> s

Textual representation.

Since: unicode-grapheme-common-0.1

Errors

newtype UnsupportedUnicodeE

Exception for unsupported unicode version.

Since: unicode-grapheme-common-0.1

Constructors

MkUnsupportedUnicodeE Version